backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "ResNex AI backend services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.50.0",
    "tiktoken>=0.7.0",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.0.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "rq>=2.6.0",
    "sqlalchemy>=2.0.43",
    "psycopg[binary]>=3.2.10",
    "pgvector>=0.3.0",
    "uvicorn[standard]>=0.41.0",
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
